First implementation of generator
See `--help` for options (which are mostly print and write currently). It expects the benchmarks.yaml in the same directory.

The script goes through the YAML and first prints the suite, then goes through all the benchmarks of the suites, and finally prints the standalone benchmarks.

Most of the setup logic is in Python, and not in Jinja, since I want to reuse the same template for suites and benchmarks. I think it's a nice way.

There's still things to be done (beyond bugfixes), most prominently the vertical line. But this is mostly esthetical, and I will add it later.